# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ujima SACCO AI Pride Ecosystem API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    get:
      operationId: getProfile
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/change-password:
    post:
      operationId: changePassword
      tags: [users]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── MEMBERS ───────────────────────────────────────────────────────────────
...
  /loan-products/{id}:
    get:
      operationId: getLoanProduct
      tags: [loans]
      summary: Get a loan product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Loan product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanProduct"

  # ── LOAN APPLICATIONS ─────────────────────────────────────────────────────
  /loan-applications:
    get:
      operationId: listLoanApplications
      tags: [loans]
      summary: List loan applications for current member
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LoanApplication"
    post:
      operationId: createLoanApplication
      tags: [loans]
      summary: Submit a loan application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanApplication"
  /loan-applications/{id}:
    get:
      operationId: getLoanApplication
      tags: [loans]
      summary: Get a loan application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanApplication"
  /loan-applications/{id}/status:
    get:
      operationId: getLoanApplicationStatus
      tags: [loans]
      summary: Get application AI pipeline status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pipeline status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationPipelineStatus"

  # ── ACTIVE LOANS ──────────────────────────────────────────────────────────
...
  /loans/{id}:
    get:
      operationId: getActiveLoan
      tags: [loans]
      summary: Get active loan detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Active loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActiveLoan"

  # ── DOCUMENTS ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── AUTH SCHEMAS ────────────────────────────────────────────────────────
...
    UserProfile:
      type: object
      required: [id, email, firstName, lastName, role]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        mfaEnabled:
          type: ["boolean", "null"]
        kycStatus:
          type: ["string", "null"]
        createdAt:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
...
    LoanProduct:
      type: object
      required: [id, name, type, minAmount, maxAmount, minTenureMonths, maxTenureMonths, interestRate]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [business, vendor, agricultural, education, emergency, asset_financing]
        description:
          type: ["string", "null"]
        minAmount:
          type: number
        maxAmount:
          type: number
        minTenureMonths:
          type: integer
        maxTenureMonths:
          type: integer
        interestRate:
          type: number
        requirements:
          type: ["string", "null"]
        isActive:
          type: boolean

    # ── LOAN APPLICATION SCHEMAS ─────────────────────────────────────────────
    LoanApplicationInput:
      type: object
      required: [loanProductId, requestedAmount, tenureMonths, purpose, employmentStatus, monthlyIncome]
      properties:
        loanProductId:
          type: integer
        requestedAmount:
          type: number
        tenureMonths:
          type: integer
        purpose:
          type: string
        employmentStatus:
          type: string
        employerName:
          type: ["string", "null"]
        monthlyIncome:
          type: number
        otherIncome:
          type: ["number", "null"]
        existingLoans:
          type: ["number", "null"]
        businessName:
          type: ["string", "null"]
        businessType:
          type: ["string", "null"]
        businessRevenue:
          type: ["number", "null"]
        collateralDescription:
          type: ["string", "null"]
    LoanApplication:
      type: object
      required: [id, userId, loanProductId, requestedAmount, tenureMonths, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        loanProductId:
          type: integer
        loanProduct:
          $ref: "#/components/schemas/LoanProduct"
        requestedAmount:
          type: number
        tenureMonths:
          type: integer
        purpose:
          type: string
        status:
          type: string
          enum: [submitted, under_review, additional_info_required, approved, rejected, disbursed]
        employmentStatus:
          type: string
        monthlyIncome:
          type: number
        approvedAmount:
          type: ["number", "null"]
        approvedTenure:
          type: ["integer", "null"]
        adminNotes:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        pipelineStage:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ApplicationPipelineStatus:
      type: object
      required: [applicationId, currentStage, stages]
      properties:
        applicationId:
          type: integer
        currentStage:
          type: string
        stages:
          type: array
          items:
            $ref: "#/components/schemas/PipelineStage"
        aiAssessment:
          $ref: "#/components/schemas/AiAssessment"
    PipelineStage:
      type: object
      required: [name, label, status]
      properties:
        name:
          type: string
        label:
          type: string
        labelSw:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, failed]
        completedAt:
          type: ["string", "null"]
...
    ActiveLoan:
      type: object
      required: [id, userId, applicationId, principalAmount, outstandingBalance, status, disbursedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        applicationId:
          type: integer
        loanProduct:
          $ref: "#/components/schemas/LoanProduct"
        principalAmount:
          type: number
        outstandingBalance:
          type: number
        interestRate:
          type: number
        tenureMonths:
          type: integer
        monthlyInstalment:
          type: number
        totalRepaid:
          type: number
        nextPaymentDate:
          type: ["string", "null"]
        maturityDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, closed, defaulted, written_off]
        disbursedAt:
          type: string
        memberName:
          type: ["string", "null"]

    # ── DOCUMENT SCHEMAS ────────────────────────────────────────────────────
...
    AiAssessment:
      type: object
      required: [id, applicationId, creditScore, riskScore, confidenceScore, approvalProbability, recommendedLimit]
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        creditScore:
          type: number
          description: 300-850
        riskScore:
          type: number
          description: 0-100 (lower is better)
        confidenceScore:
          type: number
          description: 0-100
        approvalProbability:
          type: number
          description: 0-100
        recommendedLimit:
          type: number
        riskLevel:
          type: string
          enum: [low, medium, high, very_high]
        scoutSummary:
          type: ["string", "null"]
        guardianSummary:
          type: ["string", "null"]
        hunterSummary:
          type: ["string", "null"]
        loanTerms:
          type: ["object", "null"]
        repaymentPlan:
          type: ["object", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string

    # ── DASHBOARD SCHEMAS ────────────────────────────────────────────────────
...